mutants/tests/requirements-test.txt
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-xdist==3.3.1
pytest-rerunfailures==15.0
coverage==7.3.1
python-dotenv==1.0.0 